Code refactor. Extract eventlisteners from numbers/operators functions
Separated the logic of the click eventlistener from the logic that updates the variables after a click.
Created two new functions handleNumbers and handleOperators that take the user inputs from the eventlistener as an argument, update the variables and run the calculator logic.
The new events now listen to the user inputs and call the corresponding functions